requirement.txt
asttokens==2.4.1
backcall==0.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
executing==2.0.1
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
gdown==5.2.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imgviz==1.7.5
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
labelImg==1.8.6
labelme==5.5.0
lapx==0.5.9.post1
lazy_loader==0.4
lxml==5.2.2
malaria==1.0.2
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numpy==1.24.4
object-detection-metrics==0.4.post1
onnx==1.12.0
onnxruntime==1.18.1
opencv-python==4.9.0.80
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
podm==0.0.19
prompt_toolkit==3.0.47
protobuf==3.20.1
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycocotools==2.0.7
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
qt-material==2.14
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.4.1
requests==2.32.3
sahi==0.11.18
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==70.1.1
shapely==2.0.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tensorrt==8.4.2.4
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==1.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
zipp==3.19.2
